feat: modularize Now Playing section into a new component
- Create \`src/components/now-playing.tsx\` to encapsulate the Spotify card section.
- Refactor \`src/pages/home.tsx\` to use the new \`NowPlaying\` component.
- Maintain responsive \`max-w-md\` width and consistent layout patterns.
- Ensure proper accessibility and unique SVG filter IDs.
- Pass \`pnpm check\` with no errors.
